requirements.txt
Django==3.2.4
djangorestframework==3.12.4
python-dateutil==2.8.1
loguru==0.5.3
cx-Oracle==8.2.1
drf-spectacular==0.17.0
django-cors-headers==3.7.0
requests==2.25.1
qrcode==7.2
lxml==4.6.3
flatdict==4.0.1
pre-commit==2.13.0
python-magic==0.4.24